+++ src/types.ts
@@ -42,6 +42,7 @@ export interface ExtraCommandLineOptions {
   exitWithCount?: boolean;
   excludeDeclarationFiles?: boolean;
   ignoreFilesRegex?: string[];
+  maxIssues: number;
   pathsToExcludeFromReport?: string[];
   searchNamespaces?: boolean;
   showLineNumber?: boolean;