+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Added
+
+- Add options to allow maximum number of issues: --maxIssues
+
 ## [6.0.0] - 24 Jan 2020
 
 ### Added